requirements.txt
# AWS SDK and Services
boto3>=1.34.0
botocore>=1.34.0

# AWS Bedrock Agents (Strands SDK)
strands-agents>=0.1.0
# OpenTelemetry dependencies for Strands SDK
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-propagator-b3==1.40.0
opentelemetry-propagator-jaeger==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0

# Data Validation and Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Twilio WhatsApp Integration
twilio>=8.10.0

# Google Calendar API
google-api-python-client>=2.110.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Microsoft Graph API
msal>=1.26.0
requests>=2.31.0

# Utilities
python-dateutil>=2.8.2
bleach>=6.1.0

# Logging
structlog>=23.2.0

# Web Framework (for local development)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
